Disconnected Max Endstops and Servos 2-4. Added SD card Headers. Fixed some Silkscreen errors.
This is a step towards making the LCD screen work, but in the meantime,
the last level shifter was changed to all NC. At some point, that will
be the encoder button, encoder 1, encoder 2, MISO, and whatever else
needs to be shifted to make the reprap full graphics controller work.
The issue is the documentation sucks on that so I need to really dig to
make it work. In the meantime, I added just a breakout header for the
generic 6-pin SD card module for arduinos. This doesnt replace an LCD,
but at least makes it run standalone.

The Max Endstops were not rerouted to anywhere, and are disconnected
from the rest of the system to prevent someone accidently frying their
Due. The Servo pins are outputs, so they were connected back the their
respective pins, but are not level shifted, so they may behave
strangely. The first Servo pin (SERVO0) is level shifted.